move pricer to separate module
