+++ runtime/near-evm-runner/src/lib.rs
@@ -24,6 +24,7 @@ mod builtins;
 mod evm_state;
 mod interpreter;
 mod near_ext;
+mod pricer;
 pub mod types;
 pub mod utils;
